requirements.txt
# Core dependencies
python-dotenv>=1.0.0
fastmcp>=0.1.0
fastapi>=0.68.0
uvicorn>=0.15.0
pydantic>=1.8.0

# Development dependencies
pytest>=7.0.0
black>=23.0.0
flake8>=6.0.0
mypy>=1.0.0 